+++ R/fn_Overview_data.R
@@ -1,14 +1,15 @@
 #' Function to generate data dictionary of a data frame
 #'
-#' This function used to produce summary of data frame.
+#' This function used to produce the metadata information and data summary
 #'
 #' @param data a data frame
 #' @param type Type 1 is overall data summary; Type 2 is variable level summary
+#' @param fun to add any additional statistics into metadata type 2 output, for example: mean, sum, etc..
 #' @details
-#' This function provides overall and variable level data summary like percentage of missings, variable types etc..
+#' This function provides overall and variable level data summary like percentage of missing, variable types etc..
 #' \itemize{
 #' \item Type = 1, overall data summary (column names are "Descriptions Value")
-#' \item Type = 2, variable level summary (column names are "Index Variable_Name  Variable_Type Per_of_Missing No_of_distinct_values")
+#' \item Type = 2, variable level summary (column names are "Index Variable_Name Variable_Type Sample_n Missing_count Per_of_Missing No_of_distinct_values" and other statistics)
 #'}
 #' @examples
 #' # Overall data summary
@@ -17,7 +18,7 @@
 #' ExpData(data=mtcars,type=2)
 #' @export ExpData
 
-ExpData <- function(data, type = 1){
+ExpData <- function(data, type = 1, fun=NULL){
   if (!is.data.frame(data)) stop("data must be a numeric vector or data.frame")
   xx <- as.data.frame(data)
   dd <- sapply( sapply(xx, function(x){
@@ -72,31 +73,28 @@ ExpData <- function(data, type = 1){
             ## Data Structure
             name_var <- names(xx)
             tt <- sapply(name_var, function(x){
-              Xvar <- xx[, x]
-              cla_var <- as.character(paste0(class(xx[, x]), collapse = ":"))
-              #Per_missing <- round((length(Xvar[is.na(Xvar)]) + length(Xvar[Xvar=='']))/ length(Xvar), 5)
-
-              if(is.character(Xvar)) {
-                Per_missing <- round((length(Xvar[is.na(Xvar)]) + length(Xvar[Xvar == '']))/ length(Xvar), 5)
-              } else {
-                Per_missing <- round((length(Xvar[is.na(Xvar)]))/ length(Xvar), 5)
-              }
-
-              Per_Unique <- length(unique(Xvar))
-              mydata <- c(Index = 1, VarName = x, VarClass = cla_var, Per_mis = Per_missing,
-                          Unique = Per_Unique, VarDescriptions = NULL)
-              return(mydata)
+              expdatatype2(xx, x, myfun=fun)
             }
             )
 
             op <- data.frame(t(tt))
             op$Index <- seq(1, length(name_var), 1)
             rownames(op) <- NULL
-            names(op) <- c("Index", "Variable_Name", "Variable_Type", "Per_of_Missing", "No_of_distinct_values")
             op[, 2] <- as.character(paste0(op[, 2]))
             op[, 3] <- as.character(paste0(op[, 3]))
-            op[, 4] <- as.numeric(paste0(op[, 4]))
+            op[, 4] <- as.integer(paste0(op[, 4]))
             op[, 5] <- as.integer(paste0(op[, 5]))
+            op[, 6] <- as.numeric(paste0(op[, 6]))
+            op[, 7] <- as.integer(paste0(op[, 7]))
+
+            if(!is.null(fun)) {
+              for(j in 1:length(fun)) op[, 7+j] <- as.numeric(paste0(op[, 7+j]))
+            }
+
+            names(op) <- c("Index", "Variable_Name", "Variable_Type","Sample_n","Missing_Count", "Per_of_Missing", "No_of_distinct_values", fun)
+
+            if(!is.null(fun)) op[!op$Variable_Type %in% c("integer", "numeric"), fun] <- NA
+
             return(op)
         }
-}
+ }